+++ lca-ai-stack/source/lambda_functions/call_event_stream_processor/graphql_helpers/__init__.py
@@ -1,8 +1,0 @@
-# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-# SPDX-License-Identifier: Apache-2.0
-"""GraphQL Helpers"""
-
-from .call_fields import call_fields
-from .transcript_segment_fields import transcript_segment_fields
-
-__all__ = ["call_fields", "transcript_segment_fields"]