+++ lca-ai-stack/source/lambda_layers/transcript_enrichment_layer/lex_utils/__init__.py
@@ -1,0 +1,6 @@
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0
+"""Async Lex Client Utilities"""
+from .lex import recognize_text_lex
+
+__all__ = ["recognize_text_lex"]