+++ lca-genesys-audiohook-stack/source/app/src/lca/lca.ts
@@ -12,50 +12,57 @@
 // # See the License for the specific language governing permissions and
 // # limitations under the License.
 
-import { CallEvent, CallEventStatus, CallRecordingEvent } from './entities-lca';
-import {
-    DynamoDBClient,
-    PutItemCommand
-} from '@aws-sdk/client-dynamodb';
+import { 
+    CallEvent, 
+    CallEventStatus, 
+    CallRecordingEvent 
+} from './entities-lca';
+
 import { TranscriptEvent } from '@aws-sdk/client-transcribe-streaming';
+import { 
+    KinesisClient, 
+    PutRecordCommand 
+} from '@aws-sdk/client-kinesis';
+
 import dotenv from 'dotenv';
 dotenv.config();
 
 const awsRegion:string = process.env['AWS_REGION'] || 'us-east-1';
-const ddbTableName = process.env['EVENT_SOURCING_TABLE_NAME'] || null;
-const expireInDays = Number(process.env['EXPIRATION_IN_DAYS']) || 90;
+const expireInDays = 90;
 const savePartial = process.env['SAVE_PARTIAL_TRANSCRIPTS'] === 'true';
+const kdsStreamName = process.env['KINESIS_STREAM_NAME'] || '';
 
-const dynamoClient = new DynamoDBClient({ region: awsRegion });
-export const writeCallEventToDynamo = async (callEvent: CallEvent ) => {
+const kinesisClient = new KinesisClient({ region: awsRegion });
+
+export const writeCallEventToKds = async (callEvent: CallEvent ) => {
 
     const now = new Date().toISOString();
     const expiration = Date.now() / 1000 + expireInDays * 24 * 3600;
 
+    const kdsObj =  {
+        CallId: callEvent.callId,
+        EventType: callEvent.eventStatus,
+        Channel: callEvent.channel,
+        CustomerPhoneNumber: callEvent.fromNumber || '',
+        SystemPhoneNumber: callEvent.toNumber || '',
+        CreatedAt: now,
+        ExpiresAfter: expiration.toString(),
+    };
+    
     const putParams = {
-        TableName: ddbTableName ?? '',
-        Item: {
-            PK: { 'S' : `ce#${callEvent.callId}` },
-            SK: { 'S' : `ts#${now}#et#${callEvent.eventStatus}#c#${callEvent.channel}` },
-            CallId: { 'S' : callEvent.callId },
-            EventType: { 'S' : callEvent.eventStatus },
-            Channel: { 'S' : callEvent.channel },
-            CustomerPhoneNumber: { 'S' : callEvent.fromNumber || '' },
-            SystemPhoneNumber: { 'S' : callEvent.toNumber || '' },
-            CreatedAt: { 'S' : now },
-            ExpiresAfter: { 'N' : expiration.toString() },
-        }
+        StreamName: kdsStreamName,
+        PartitionKey: callEvent.callId,
+        Data: Buffer.from(JSON.stringify(kdsObj))
     };
 
-    const putCmd = new PutItemCommand(putParams);
+    const putCmd = new PutRecordCommand(putParams);
     try {
-        await dynamoClient.send(putCmd);
-    } catch (err) {
-        console.error(err);
+        await kinesisClient.send(putCmd);
+    } catch (error) {
+        console.error('Error writing transcription segment to KDS', error);
     }
 };
-
-export const writeRecordingUrlToDynamo = async (recordingEvent: CallRecordingEvent) => {
+export const writeRecordingUrlToKds = async (recordingEvent: CallRecordingEvent) => {
 
     const url = new URL(recordingEvent.recordingsKeyPrefix+recordingEvent.recordingsKey, `https://${recordingEvent.recordingsBucket}.s3.${awsRegion}.amazonaws.com`);
     const recordingUrl = url.href;
@@ -64,53 +71,54 @@ export const writeRecordingUrlToDynamo = async (recordingEvent: CallRecordingEve
     const currentTimeStamp = now.toISOString();
     const expiresAfter = Math.ceil((Number(now) + expireInDays * 24 * 3600 * 1000) / 1000,);
   
-    const putParams = {
-        TableName: ddbTableName ?? '',
-        Item : {
-            PK: { 'S' : `ce#${recordingEvent.callId}` },  
-            SK: { 'S' : `ts#${currentTimeStamp}#et#${recordingEvent.eventType}` },
-            CallId: { 'S' : recordingEvent.callId },
-            ExpiresAfter: { 'N' : expiresAfter.toString() },
-            CreatedAt: { 'S' : currentTimeStamp },
-            RecordingUrl: { 'S' : recordingUrl },
-            EventType: { 'S' : recordingEvent.eventType },
-        }
+    const kdsObj =  {
+        CallId: recordingEvent.callId,
+        ExpiresAfter: expiresAfter.toString(),
+        CreatedAt: currentTimeStamp,
+        RecordingUrl: recordingUrl,
+        EventType: recordingEvent.eventType,
     };
-    
-    const putCmd = new PutItemCommand(putParams);
-  
+
+    const putParams = {
+        StreamName: kdsStreamName,
+        PartitionKey: recordingEvent.callId,
+        Data: Buffer.from(JSON.stringify(kdsObj))
+    };
+
+    const putCmd = new PutRecordCommand(putParams);
     try {
-        await dynamoClient.send(putCmd);
+        await kinesisClient.send(putCmd);
     } catch (error) {
-        console.error(error);  
+        console.error('Error writing transcription segment to KDS', error);
     }
 };
 
-export const writeStatusToDynamo = async (status: CallEventStatus) => {
+export const writeStatusToKds = async (status: CallEventStatus) => {
 
     const now = new Date().toISOString();
     const expiration = Date.now() / 1000 + expireInDays * 24 * 3600;
   
+    const kdsObj =  {
+        CallId: status.callId,
+        EventType: status.eventStatus,
+        Channel: status.channel,
+        TransactionId: status.transactionId || '', 
+        CreatedAt: now,
+        ExpiresAfter: expiration.toString(),
+    };
     const putParams = {
-        TableName: ddbTableName ?? '',
-        Item: {
-            PK: { 'S' : `ce#${status.callId}` },
-            SK: { 'S' : `ts#${now}#et${status.eventStatus}#c#${status.channel}` },
-            CallId: { 'S' : status.callId },
-            EventType: { 'S' : status.eventStatus },
-            Channel: { 'S' : status.channel },
-            TransactionId: { 'S': status.transactionId || '' }, 
-            CreatedAt: { 'S' : now },
-            ExpiresAfter: { 'N' : expiration.toString() },
-        }
+        StreamName: kdsStreamName,
+        PartitionKey: status.callId,
+        Data: Buffer.from(JSON.stringify(kdsObj))
     };
 
-    const putCmd = new PutItemCommand(putParams);
+    const putCmd = new PutRecordCommand(putParams);
     try {
-        await dynamoClient.send(putCmd);
-    } catch (err) {
-        console.error(err);
+        await kinesisClient.send(putCmd);
+    } catch (error) {
+        console.error('Error writing transcription segment to KDS', error);
     }
+
 };
 
 export const writeTranscriptionSegment = async function(transcribeMessageJson:TranscriptEvent, callId: string, transactionId:string | undefined) {
@@ -135,30 +143,32 @@ export const writeTranscriptionSegment = async function(transcribeMessageJson:Tr
             const now = new Date().toISOString();
             const expiration = Math.round(Date.now() / 1000) + expireInDays * 24 * 3600;
             const eventType = 'ADD_TRANSCRIPT_SEGMENT';
-            const putParams = {
-                TableName: ddbTableName ?? '',
-                Item : {
-                    PK: { 'S' : `ce#${callId}` },
-                    SK: { 'S' : `ts#${now}#et#${eventType}#c#${channel}` },
-                    Channel: { 'S' : channel },
-                    TransactionId: { 'S': transid },
-                    CallId: { 'S': callId },
-                    SegmentId: { 'S': resultId },
-                    StartTime: { 'N': startTime.toString() },
-                    EndTime: { 'N': endTime.toString() },
-                    Transcript: { 'S': transcript || '' },
-                    IsPartial: { 'BOOL': ispartial },
-                    EventType: { 'S': eventType.toString() },
-                    CreatedAt: { 'S': now },
-                    ExpiresAfter: { 'N': expiration.toString() }
-                }
+            const kdsObject = {
+                Channel: channel,
+                TransactionId: transid,
+                CallId: callId,
+                SegmentId: resultId,
+                StartTime: startTime.toString(),
+                EndTime: endTime.toString(),
+                Transcript: transcript || '',
+                IsPartial: ispartial,
+                EventType: eventType.toString(),
+                CreatedAt: now,
+                ExpiresAfter: expiration.toString(),
+                StreamArn: '' 
             };
             
-            const putCmd = new PutItemCommand(putParams);
+            const putParams = {
+                StreamName: kdsStreamName,
+                PartitionKey: callId,
+                Data: Buffer.from(JSON.stringify(kdsObject)),
+            };
+
+            const putCmd = new PutRecordCommand(putParams);
             try {
-                await dynamoClient.send(putCmd);
+                await kinesisClient.send(putCmd);
             } catch (error) {
-                console.error(error);
+                console.error('Error writing transcription segment to KDS', error);
             }
         }
     }